Create main.py
(Only works on linux) This is the main program file for ScorPyon. It should be in the same directory as "runner.py" to work if you do not want to use the install scripts. Go inside and just change the "os.chdir" lines to the directory where you have kept the file.