Update main.py
(only works on linux) This is the main program file for ScorPyon and should always be in the same directory as "runner.py". If you do not want to use the installation scripts, just go inside and change the "os.chdir" path at the begginging of the file to wherever it actually is.