Create main.py
(only works on linux) This is the main program file for ScorPyon. It should always be in the same directory as "runner.py" to work. If you do not want to use the installation scripts, make sure the "os.chdir" at the beggining of the file sets the correct path and that exists.