Create main.py
(only works on linux) This is the main file for ScorPyon. It should always be in the same directory as runner.py for the program to work. If you do not want to use the installation scripts, make sure the "os.chdir" command at the beginning of the file is pointing towards the correct path.